ADD : Moved slatekit.core.app into its own project
