+++ src/apps/kotlin/slatekit-sampleapps/sampleapp-core/settings.gradle
@@ -3,6 +3,9 @@ rootProject.name = 'slatekit-sampleapp-core'
 include ':slatekit-common'
 project(':slatekit-common').projectDir=new File('../../../../lib/kotlin/slatekit-common')
 
+include ':slatekit-app'
+project(':slatekit-app').projectDir=new File('../../../../lib/kotlin/slatekit-app')
+
 include ':slatekit-meta'
 project(':slatekit-meta').projectDir=new File('../../../../lib/kotlin/slatekit-meta')
