+++ src/apps/kotlin/slatekit-sampleapps/sampleapp-server/settings.gradle
@@ -3,6 +3,9 @@ rootProject.name = 'slatekit-sampleapp-server'
 include ':slatekit-common'
 project(':slatekit-common').projectDir=new File('../../../../lib/kotlin/slatekit-common')
 
+include ':slatekit-app'
+project(':slatekit-app').projectDir=new File('../../../../lib/kotlin/slatekit-app')
+
 include ':slatekit-meta'
 project(':slatekit-meta').projectDir=new File('../../../../lib/kotlin/slatekit-meta')
