+++ src/lib/kotlin/settings.gradle
@@ -1,4 +1,5 @@
 include 'slatekit-common',
+        'slatekit-app',
         'slatekit-db',
         'slatekit-meta',
         'slatekit-entities',
