+++ src/lib/kotlin/slatekit-app/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Mon Oct 26 12:25:13 CET 2015
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=http\://services.gradle.org/distributions/gradle-2.7-all.zip
