+++ src/lib/kotlin/slatekit-app/settings.gradle
@@ -1,0 +1,4 @@
+rootProject.name = 'slatekit-app'
+
+include ':slatekit-common'
+//project(':slatekit-common').projectDir=new File('../slatekit-common')
