+++ src/lib/kotlin/slatekit-examples/build.gradle
@@ -51,6 +51,7 @@ dependencies {
     compile "com.googlecode.json-simple:json-simple:1.1"
     compile 'com.slatekit:slatekit-results:0.9.8'
     compile project(":slatekit-common")
+    compile project(":slatekit-app")
     compile project(":slatekit-meta")
     compile project(":slatekit-entities")
     compile project(":slatekit-core")
