+++ src/lib/kotlin/slatekit-integration/src/main/kotlin/slatekit/integration/common/AppEntContext.kt
@@ -28,7 +28,6 @@ import slatekit.common.info.Status
 import slatekit.common.log.Logs
 import slatekit.common.log.LogsDefault
 import slatekit.common.naming.Namer
-import slatekit.core.app.AppBuilder
 import slatekit.core.common.AppContext
 import slatekit.entities.core.Entities
 
@@ -84,7 +83,7 @@ data class AppEntContext(
          * the same context without the Entities
          */
         fun fromContext(ctx: Context, namer: Namer? = null): AppEntContext {
-            val dbCons = AppBuilder.dbs(ctx.cfg)
+            val dbCons = DbLookup.fromConfig(ctx.cfg)
             return AppEntContext(
                     ctx.arg, ctx.env, ctx.cfg, ctx.logs, ctx.app, ctx.sys, ctx.build, ctx.start, Entities(dbCons, ctx.enc, namer = namer), dbCons, ctx.enc, ctx.dirs
             )
@@ -96,7 +95,7 @@ data class AppEntContext(
          * the same context without the Entities
          */
         fun fromAppContext(ctx: AppContext, namer: Namer? = null): AppEntContext {
-            val dbCons = AppBuilder.dbs(ctx.cfg)
+            val dbCons = DbLookup.fromConfig(ctx.cfg)
             return AppEntContext(
                     ctx.arg, ctx.env, ctx.cfg, ctx.logs, ctx.app, ctx.sys, ctx.build, ctx.start, Entities(dbCons, ctx.enc, namer = namer), dbCons, ctx.enc, ctx.dirs
             )
