+++ src/lib/kotlin/slatekit-tests/build.gradle
@@ -43,6 +43,7 @@ dependencies {
     compile "postgresql:postgresql:9.1-901-1.jdbc4"
     compile 'com.slatekit:slatekit-results:0.9.8'
     compile project(":slatekit-common")
+    compile project(":slatekit-app")
     compile project(":slatekit-db")
     compile project(":slatekit-query")
     compile project(":slatekit-meta")
