+++ src/lib/kotlin/slatekit-tests/settings.gradle
@@ -3,6 +3,9 @@ rootProject.name = 'slatekit-tests'
 include ':slatekit-common'
 project(':slatekit-common').projectDir=new File('../slatekit-common')
 
+include ':slatekit-app'
+project(':slatekit-app').projectDir=new File('../../../../lib/kotlin/slatekit-app')
+
 include ':slatekit-meta'
 project(':slatekit-meta').projectDir=new File('../slatekit-meta')
