+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/app/AppMetaTests.kt
@@ -14,16 +14,9 @@ package slate.test
 import org.junit.Assert
 import org.junit.Test
 import slatekit.common.*
-import slatekit.common.args.Arg
-import slatekit.common.args.ArgsSchema
 import slatekit.common.info.About
-import slatekit.core.app.App
-import slatekit.core.app.AppRunner
-import slatekit.core.common.AppContext
-import slatekit.results.StatusCodes
-import slatekit.results.Success
-import slatekit.results.Try
-import slatekit.results.getOrElse
+import slatekit.app.App
+import slatekit.app.AppRunner
 
 
 class AppMetaTests  {
